Refactor LocalCorrectionsSubProcessor to jdt.core.manipulation
- create new LocalCorrectionsBaseSubProcessor in jdt.core.manipulation
  and have LocalCorrectionsSubProcessor sub-class it
- add new constructors where needed to support specifying a core
  class and an image
- fixes #1847
