+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/text/correction/QuickAssistProcessor.java
@@ -1858,28 +1858,6 @@ public class QuickAssistProcessor implements IQuickAssistProcessor {
 		return false;
 	}
 
-	public static ASTNode getCopyOfInner(ASTRewrite rewrite, ASTNode statement, boolean toControlStatementBody) {
-		if (statement.getNodeType() == ASTNode.BLOCK) {
-			Block block= (Block) statement;
-			List<Statement> innerStatements= block.statements();
-			int nStatements= innerStatements.size();
-			if (nStatements == 1) {
-				return rewrite.createCopyTarget(innerStatements.get(0));
-			} else if (nStatements > 1) {
-				if (toControlStatementBody) {
-					return rewrite.createCopyTarget(block);
-				}
-				ListRewrite listRewrite= rewrite.getListRewrite(block, Block.STATEMENTS_PROPERTY);
-				ASTNode first= innerStatements.get(0);
-				ASTNode last= innerStatements.get(nStatements - 1);
-				return listRewrite.createCopyTarget(first, last);
-			}
-			return null;
-		} else {
-			return rewrite.createCopyTarget(statement);
-		}
-	}
-
 
 	private static boolean getUnWrapProposals(IInvocationContext context, ASTNode node, Collection<ICommandAccess> resultingCollections) {
 		ASTNode outer= node;
@@ -1975,7 +1953,7 @@ public class QuickAssistProcessor implements IQuickAssistProcessor {
 			return false;
 		}
 		ASTRewrite rewrite= ASTRewrite.create(outer.getAST());
-		ASTNode inner= getCopyOfInner(rewrite, body, ASTNodes.isControlStatementBody(outer.getLocationInParent()));
+		ASTNode inner= QuickAssistProcessorUtil.getCopyOfInner(rewrite, body, ASTNodes.isControlStatementBody(outer.getLocationInParent()));
 		if (inner == null) {
 			return false;
 		}
