+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/text/correction/proposals/RefactoringCorrectionProposal.java
@@ -30,4 +30,7 @@ public class RefactoringCorrectionProposal extends LinkedCorrectionProposal {
 		super(name, cu, null, relevance, image, delegate);
 	}
 
+	public RefactoringCorrectionProposal(RefactoringCorrectionProposalCore core, Image image) {
+		super(core, image);
+	}
 }
